Update file paths for component file tree restructure
