--- src/components/App/App.test.js
@@ -1,1 +1,0 @@
-
